Lifted up the game state from the child Component, Square, and placed into the parent Component, Board.
Set the Board component to declare the state variable that defaults to an array of 9 nulls (values), associated with the 9 squares.

Updated Board component to pass the value prop down to each Square component, the prop value is associated with one of the elements in the array state variable.

Updated the Square component function to take in the value prop from the Board component through its parameter.

Removed the Square component's stateful tracking of value and the button's onClick prop.